add scalefx AA and 3x, and super-xbr 2x modes
super-xbr is based on the c-port, which is based on super-xbr-2x-fast. additionally optimized and templated for (missing) alpha.
super-xbr is usually better then xbr(z) on higher res and/or higher bpp content.

scalefx is mostly the same as the old VPX code.
scalefx is usually the best option for DMD content from all my tests.

+++ plugins/upscaledmd/xbrz/xbrz.h
@@ -85,8 +85,8 @@ inline
 void nearestNeighborScale(const uint32_t* src, int srcWidth, int srcHeight,
                           uint32_t* trg, int trgWidth, int trgHeight)
 {
-    nearestNeighborScale(src, srcWidth, srcHeight, srcWidth * sizeof(uint32_t),
-                         trg, trgWidth, trgHeight, trgWidth * sizeof(uint32_t),
+    nearestNeighborScale(src, srcWidth, srcHeight, srcWidth * (int)sizeof(uint32_t),
+                         trg, trgWidth, trgHeight, trgWidth * (int)sizeof(uint32_t),
                          NN_SCALE_SLICE_TARGET, 0, trgHeight);
 }
 }
